STSR-2025/baseline/STSR-Task2-Baseline/requirement.txt
numpy
torch
torchvision
torchaudio
nibabel
trimesh
open3d
PyYAML
tqdm
scipy
tensorboard
pytorch_geometric